backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1
email-validator==2.1.0.post1
httpx==0.26.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
typing-extensions>=4.0.0
fastapi-mail==1.4.1
reportlab==4.0.7
razorpay==1.4.1